Move tool source cache to tools.cache
