+++ lib/galaxy/tools/cache.py
@@ -1,18 +1,92 @@
+import json
 import logging
 import os
 from collections import defaultdict
 from threading import Lock
 
+from dogpile.cache import make_region
+from dogpile.cache.api import (
+    CachedValue,
+    NO_VALUE,
+)
+from dogpile.cache.backends.file import AbstractFileLock
+from dogpile.cache.proxy import ProxyBackend
+from dogpile.util import ReadWriteMutex
+from lxml import etree
 from sqlalchemy.orm import (
     defer,
     joinedload,
 )
 
-from galaxy.tools import region
+from galaxy.tool_util.parser import get_tool_source
 from galaxy.util import unicodify
 from galaxy.util.hash_util import md5_hash_file
 
 log = logging.getLogger(__name__)
+
+
+class JSONBackend(ProxyBackend):
+
+    def set(self, key, value):
+        with self.proxied._dbm_file(True) as dbm:
+            dbm[key] = json.dumps({'metadata': value.metadata, 'payload': self.value_encode(value), 'macro_paths': value.payload.macro_paths()})
+
+    def get(self, key):
+        with self.proxied._dbm_file(False) as dbm:
+            if hasattr(dbm, "get"):
+                value = dbm.get(key, NO_VALUE)
+            else:
+                # gdbm objects lack a .get method
+                try:
+                    value = dbm[key]
+                except KeyError:
+                    value = NO_VALUE
+            if value is not NO_VALUE:
+                value = self.value_decode(value)
+            return value
+
+    def value_decode(self, v):
+        if not v or v is NO_VALUE:
+            return NO_VALUE
+        # v is returned as bytestring, so we need to `unicodify` on python < 3.6 before we can use json.loads
+        v = json.loads(unicodify(v))
+        payload = get_tool_source(xml_tree=etree.ElementTree(etree.fromstring(v['payload'].encode('utf-8'))), macro_paths=v['macro_paths'])
+        return CachedValue(metadata=v['metadata'], payload=payload)
+
+    def value_encode(self, v):
+        return unicodify(v.payload.to_string())
+
+
+class MutexLock(AbstractFileLock):
+    def __init__(self, filename):
+        self.mutex = ReadWriteMutex()
+
+    def acquire_read_lock(self, wait):
+        # No need for read lock. It is supposed to prevent the "dogpile" effect
+        # where multiple functions each create the cached resource, but I don't
+        # think we care.
+        return True
+
+    def acquire_write_lock(self, wait):
+        ret = self.mutex.acquire_write_lock(wait)
+        return wait or ret
+
+    def release_read_lock(self):
+        return True
+
+    def release_write_lock(self):
+        return self.mutex.release_write_lock()
+
+
+def my_key_generator(namespace, fn, **kw):
+
+    def generate_key(*arg):
+        return "_".join(str(s) for s in arg if isinstance(s, str))
+
+    return generate_key
+
+
+region = make_region(function_key_generator=my_key_generator)
 
 
 class ToolCache(object):
